Clean up of package `stream`, pt2 (move package org.frankframework.stream.document to org.frankframework.documentbuilder)
