+++ core/src/main/java/org/frankframework/documentbuilder/DocumentFormat.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.stream.document;
+package org.frankframework.documentbuilder;
 
 public enum DocumentFormat {
 	XML,
