+++ core/src/main/java/org/frankframework/documentbuilder/IDocumentBuilder.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2021-2023 WeAreFrank!
+   Copyright 2021 WeAreFrank!
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -13,20 +13,13 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.stream.document;
+package org.frankframework.documentbuilder;
 
 import org.xml.sax.SAXException;
 
-public interface INodeBuilder extends AutoCloseable {
+public interface IDocumentBuilder extends INodeBuilder {
 
-	public ArrayBuilder startArray(String elementName) throws SAXException;
-	public ObjectBuilder startObject() throws SAXException;
-	public void setValue(String value) throws SAXException;
-	public void setValue(Number value) throws SAXException;
-	public void setValue(boolean value) throws SAXException;
-	public void setNumberValue(String value) throws SAXException;
-
-	@Override
-	public void close() throws SAXException;
+	public ObjectBuilder asObjectBuilder() throws SAXException;
+	public ArrayBuilder asArrayBuilder(String elementName) throws SAXException;
 
 }
