+++ core/src/main/java/org/frankframework/documentbuilder/JsonArrayBuilder.java
@@ -1,0 +1,42 @@
+/*
+   Copyright 2021 WeAreFrank!
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+package org.frankframework.documentbuilder;
+
+import org.xml.sax.SAXException;
+
+public class JsonArrayBuilder extends ArrayBuilder {
+
+	private final JsonEventHandler handler;
+
+	public JsonArrayBuilder(JsonEventHandler handler) throws SAXException {
+		this.handler = handler;
+		handler.startArray();
+	}
+
+	@Override
+	public INodeBuilder addElement() {
+		return new JsonNodeBuilder(handler);
+	}
+
+	@Override
+	public void close() throws SAXException {
+		try {
+			handler.endArray();
+		} finally {
+			super.close();
+		}
+	}
+}
