+++ core/src/main/java/org/frankframework/documentbuilder/xml/JsonXslt3XmlReader.java
@@ -13,9 +13,9 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.stream.document.xml;
+package org.frankframework.documentbuilder.xml;
 
-import static org.frankframework.stream.document.xml.JsonXslt3XmlHandler.TARGET_NAMESPACE;
+import static org.frankframework.documentbuilder.xml.JsonXslt3XmlHandler.TARGET_NAMESPACE;
 
 import java.io.IOException;
 import java.io.InputStream;
