+++ core/src/main/java/org/frankframework/mongodb/JsonDocumentWriter.java
@@ -28,7 +28,7 @@ import org.bson.json.StrictCharacterStreamJsonWriterSettings;
 import org.bson.json.StrictJsonWriter;
 import org.bson.types.Decimal128;
 import org.bson.types.ObjectId;
-import org.frankframework.stream.document.INodeBuilder;
+import org.frankframework.documentbuilder.INodeBuilder;
 
 /**
  * BSON writer to write to FF DocumentBuilder, to generate JSON or XML.
@@ -277,4 +277,3 @@ public class JsonDocumentWriter extends AbstractBsonWriter {
 		}
 	}
 }
-
