+++ core/src/main/java/org/frankframework/pipes/Json2XmlValidator.java
@@ -24,17 +24,11 @@ import java.util.Optional;
 
 import javax.xml.validation.ValidatorHandler;
 
-import org.apache.commons.lang3.StringUtils;
-import org.apache.xerces.xs.XSModel;
-import org.springframework.http.MediaType;
-import org.springframework.util.MimeType;
-import org.springframework.util.MimeTypeUtils;
-import org.xml.sax.ContentHandler;
-import org.xml.sax.helpers.XMLFilterImpl;
-
 import jakarta.json.Json;
 import jakarta.json.JsonStructure;
 import lombok.Getter;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.xerces.xs.XSModel;
 import org.frankframework.align.Json2Xml;
 import org.frankframework.align.Xml2Json;
 import org.frankframework.align.XmlAligner;
@@ -46,11 +40,11 @@ import org.frankframework.core.PipeForward;
 import org.frankframework.core.PipeLineSession;
 import org.frankframework.core.PipeRunException;
 import org.frankframework.core.PipeRunResult;
+import org.frankframework.documentbuilder.DocumentFormat;
 import org.frankframework.parameters.ParameterList;
 import org.frankframework.stream.Message;
 import org.frankframework.stream.MessageBuilder;
 import org.frankframework.stream.MessageContext;
-import org.frankframework.stream.document.DocumentFormat;
 import org.frankframework.util.EnumUtils;
 import org.frankframework.util.StringUtil;
 import org.frankframework.util.XmlException;
@@ -62,6 +56,11 @@ import org.frankframework.validation.XmlValidatorException;
 import org.frankframework.xml.NamespaceRemovingFilter;
 import org.frankframework.xml.RootElementToSessionKeyFilter;
 import org.frankframework.xml.XmlWriter;
+import org.springframework.http.MediaType;
+import org.springframework.util.MimeType;
+import org.springframework.util.MimeTypeUtils;
+import org.xml.sax.ContentHandler;
+import org.xml.sax.helpers.XMLFilterImpl;
 
 /**
  *<code>Pipe</code> that validates the XML or JSON input message against a XML Schema and returns either XML or JSON.
