+++ core/src/main/java/org/frankframework/senders/JsonXsltSender.java
@@ -22,9 +22,9 @@ import lombok.Setter;
 import org.apache.commons.lang3.StringUtils;
 import org.frankframework.configuration.ConfigurationException;
 import org.frankframework.core.PipeLineSession;
+import org.frankframework.documentbuilder.xml.JsonXslt3XmlReader;
 import org.frankframework.stream.Message;
 import org.frankframework.stream.MessageBuilder;
-import org.frankframework.stream.document.xml.JsonXslt3XmlReader;
 import org.frankframework.threading.ThreadConnector;
 import org.frankframework.util.TransformerPool;
 import org.frankframework.util.XmlJsonWriter;
