+++ core/src/main/java/org/frankframework/senders/XsltSender.java
@@ -32,6 +32,7 @@ import org.frankframework.core.ParameterException;
 import org.frankframework.core.PipeLineSession;
 import org.frankframework.core.SenderException;
 import org.frankframework.core.SenderResult;
+import org.frankframework.documentbuilder.xml.XmlTap;
 import org.frankframework.jta.IThreadConnectableTransactionManager;
 import org.frankframework.parameters.IParameter;
 import org.frankframework.parameters.ParameterList;
@@ -39,7 +40,6 @@ import org.frankframework.parameters.ParameterType;
 import org.frankframework.parameters.ParameterValueList;
 import org.frankframework.stream.Message;
 import org.frankframework.stream.MessageBuilder;
-import org.frankframework.stream.document.xml.XmlTap;
 import org.frankframework.threading.IThreadCreator;
 import org.frankframework.threading.ThreadConnector;
 import org.frankframework.threading.ThreadLifeCycleEventListener;
