+++ core/src/main/java/org/frankframework/stream/MessageBuilder.java
@@ -25,7 +25,7 @@ import java.nio.file.Path;
 
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.io.output.XmlStreamWriter;
-import org.frankframework.stream.document.json.JsonWriter;
+import org.frankframework.documentbuilder.json.JsonWriter;
 import org.frankframework.util.AppConstants;
 import org.frankframework.util.TemporaryDirectoryUtils;
 import org.frankframework.xml.XmlWriter;
