+++ core/src/main/java/org/frankframework/util/JdbcUtil.java
@@ -56,6 +56,9 @@ import org.frankframework.core.PipeLineSession;
 import org.frankframework.dbms.DbmsException;
 import org.frankframework.dbms.IDbmsSupport;
 import org.frankframework.dbms.JdbcException;
+import org.frankframework.documentbuilder.ArrayBuilder;
+import org.frankframework.documentbuilder.INodeBuilder;
+import org.frankframework.documentbuilder.ObjectBuilder;
 import org.frankframework.jms.BytesMessageInputStream;
 import org.frankframework.parameters.Parameter;
 import org.frankframework.parameters.ParameterList;
@@ -65,9 +68,6 @@ import org.frankframework.parameters.ParameterValueList;
 import org.frankframework.pipes.Base64Pipe.Direction;
 import org.frankframework.receivers.MessageWrapper;
 import org.frankframework.stream.Message;
-import org.frankframework.stream.document.ArrayBuilder;
-import org.frankframework.stream.document.INodeBuilder;
-import org.frankframework.stream.document.ObjectBuilder;
 import org.frankframework.xml.SaxElementBuilder;
 import org.xml.sax.SAXException;
