+++ core/src/test/java/org/frankframework/jdbc/FixedQuerySenderTest.java
@@ -19,11 +19,11 @@ import org.frankframework.core.ConfiguredTestBase;
 import org.frankframework.core.PipeLineSession;
 import org.frankframework.core.SenderException;
 import org.frankframework.dbms.Dbms;
+import org.frankframework.documentbuilder.DocumentFormat;
 import org.frankframework.functional.ThrowingConsumer;
 import org.frankframework.parameters.Parameter;
 import org.frankframework.parameters.ParameterType;
 import org.frankframework.stream.Message;
-import org.frankframework.stream.document.DocumentFormat;
 import org.frankframework.testutil.NumberParameterBuilder;
 import org.frankframework.testutil.ParameterBuilder;
 import org.frankframework.testutil.TestConfiguration;
