+++ core/src/test/java/org/frankframework/stream/document/DocumentUtilsTest.java
@@ -9,7 +9,10 @@ import jakarta.json.Json;
 import jakarta.json.JsonObject;
 import jakarta.json.JsonReader;
 import jakarta.json.JsonValue;
-import org.frankframework.stream.document.json.JsonUtils;
+import org.frankframework.documentbuilder.DocumentUtils;
+import org.frankframework.documentbuilder.Json2XmlHandler;
+import org.frankframework.documentbuilder.XmlDocumentBuilder;
+import org.frankframework.documentbuilder.json.JsonUtils;
 import org.frankframework.testutil.MatchUtils;
 import org.frankframework.xml.XmlWriter;
 import org.xml.sax.SAXException;
