+++ core/src/test/java/org/frankframework/stream/xml/XmlTapTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-import org.frankframework.stream.document.xml.XmlTap;
+import org.frankframework.documentbuilder.xml.XmlTap;
 import org.frankframework.testutil.TestFileUtils;
 import org.frankframework.util.XmlUtils;
 import org.frankframework.xml.XmlWriter;
