+++ filesystem/src/main/java/org/frankframework/filesystem/FileSystemPipe.java
@@ -32,11 +32,11 @@ import org.frankframework.core.PipeStartException;
 import org.frankframework.doc.ElementType;
 import org.frankframework.doc.ElementType.ElementTypes;
 import org.frankframework.doc.ReferTo;
+import org.frankframework.documentbuilder.DocumentFormat;
 import org.frankframework.parameters.ParameterList;
 import org.frankframework.parameters.ParameterValueList;
 import org.frankframework.pipes.FixedForwardPipe;
 import org.frankframework.stream.Message;
-import org.frankframework.stream.document.DocumentFormat;
 import org.frankframework.util.SpringUtils;
 
 /**
