+++ ladybug/src/main/java/org/frankframework/ibistesttool/IbisDebuggerAdvice.java
@@ -38,6 +38,7 @@ import org.frankframework.core.PipeLineSession;
 import org.frankframework.core.PipeRunResult;
 import org.frankframework.core.RequestReplyExecutor;
 import org.frankframework.core.SenderResult;
+import org.frankframework.documentbuilder.xml.XmlTee;
 import org.frankframework.management.bus.DebuggerStatusChangedEvent;
 import org.frankframework.parameters.Parameter;
 import org.frankframework.parameters.ParameterList;
@@ -50,7 +51,6 @@ import org.frankframework.scheduler.job.SendMessageJob.SendMessageJobSender;
 import org.frankframework.senders.ParallelSenderExecutor;
 import org.frankframework.senders.SenderWrapperBase;
 import org.frankframework.stream.Message;
-import org.frankframework.stream.document.xml.XmlTee;
 import org.frankframework.threading.ThreadConnector;
 import org.frankframework.threading.ThreadLifeCycleEventListener;
 import org.frankframework.util.AppConstants;
